requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
python-dotenv
openai
